in tagging branch will also rename read.py to read_apis.py. Hopefully no merge conflict later
